replaced when with if check

+++ Jenkinsfile
@@ -29,34 +29,29 @@ pipeline {
           },
 
           "Ruff": {
-            stage('Ruff') {   // required wrapper when using 'when' inside parallel
-              when {
-                allOf {
-                  environment name: 'CHANGE_ID', value: ''
-                  not { branch 'master' }
+            node(label: 'docker') {
+              script {
+                if (!(env.CHANGE_BRANCH != "master" && env.CHANGE_ID == '')) {
+                  return
                 }
-              }
-              node(label: 'docker') {
-                script {
-                  checkout scm
-                  fix_result = sh(script: '''docker run --pull=always --name="$BUILD_TAG-ruff-fix" -e GIT_SRC="https://github.com/eea/$GIT_NAME.git" -e GIT_NAME="$GIT_NAME" -e GIT_BRANCH="$BRANCH_NAME" -e GIT_CHANGE_ID="$CHANGE_ID" eeacms/ruff format''', returnStatus: true)
-                  sh '''docker cp $BUILD_TAG-ruff-fix:/code/$GIT_NAME .'''
-                  sh '''cp -rf eea.api.dataconnector/* .'''
-                  sh '''rm -rf eea.api.dataconnector'''
-                  sh '''docker rm -v $BUILD_TAG-ruff-fix'''
-                  FOUND_FIX = sh(script: '''git diff --name-only '*.py' | wc -l''', returnStdout: true).trim()
+                checkout scm
+                fix_result = sh(script: '''docker run --pull=always --name="$BUILD_TAG-ruff-fix" -e GIT_SRC="https://github.com/eea/$GIT_NAME.git" -e GIT_NAME="$GIT_NAME" -e GIT_BRANCH="$BRANCH_NAME" -e GIT_CHANGE_ID="$CHANGE_ID" eeacms/ruff format''', returnStatus: true)
+                sh '''docker cp $BUILD_TAG-ruff-fix:/code/$GIT_NAME .'''
+                sh '''cp -rf eea.api.dataconnector/* .'''
+                sh '''rm -rf eea.api.dataconnector'''
+                sh '''docker rm -v $BUILD_TAG-ruff-fix'''
+                FOUND_FIX = sh(script: '''git diff --name-only '*.py' | wc -l''', returnStdout: true).trim()
 
-                  if (FOUND_FIX != '0') {
-                    withCredentials([string(credentialsId: 'eea-jenkins-token', variable: 'GITHUB_TOKEN')]) {
-                      sh '''sed -i "s|url = .*|url = https://eea-jenkins:$GITHUB_TOKEN@github.com/eea/$GIT_NAME.git|" .git/config'''
-                    }
-                    sh '''git fetch origin $GIT_BRANCH:$GIT_BRANCH'''
-                    sh '''git checkout $GIT_BRANCH'''
-                    sh '''git add -- '*.py' '''
-                    sh '''git commit -m "style: Automated code fix" '''
-                    sh '''git push --set-upstream origin $GIT_BRANCH'''
-                    sh '''exit 1'''
+                if (FOUND_FIX != '0') {
+                  withCredentials([string(credentialsId: 'eea-jenkins-token', variable: 'GITHUB_TOKEN')]) {
+                    sh '''sed -i "s|url = .*|url = https://eea-jenkins:$GITHUB_TOKEN@github.com/eea/$GIT_NAME.git|" .git/config'''
                   }
+                  sh '''git fetch origin $GIT_BRANCH:$GIT_BRANCH'''
+                  sh '''git checkout $GIT_BRANCH'''
+                  sh '''git add -- '*.py' '''
+                  sh '''git commit -m "style: Automated code fix" '''
+                  sh '''git push --set-upstream origin $GIT_BRANCH'''
+                  sh '''exit 1'''
                 }
               }
             }
@@ -82,34 +77,29 @@ pipeline {
           },
 
           "Ruff": {
-            stage('Ruff') {   // required wrapper when using 'when' inside parallel
-              when {
-                allOf {
-                  environment name: 'CHANGE_ID', value: ''
-                  not { branch 'master' }
+            node(label: 'docker') {
+              script {
+                if (!(env.CHANGE_BRANCH != "master" && env.CHANGE_ID == '')) {
+                  return
                 }
-              }
-              node(label: 'docker') {
-                script {
-                  checkout scm
-                  fix_result = sh(script: '''docker run --pull=always --name="$BUILD_TAG-ruff-fix" -e GIT_SRC="https://github.com/eea/$GIT_NAME.git" -e GIT_NAME="$GIT_NAME" -e GIT_BRANCH="$BRANCH_NAME" -e GIT_CHANGE_ID="$CHANGE_ID" eeacms/ruff check''', returnStatus: true)
-                  sh '''docker cp $BUILD_TAG-ruff-fix:/code/$GIT_NAME .'''
-                  sh '''cp -rf eea.api.dataconnector/* .'''
-                  sh '''rm -rf eea.api.dataconnector'''
-                  sh '''docker rm -v $BUILD_TAG-ruff-fix'''
-                  FOUND_FIX = sh(script: '''git diff --name-only '*.py' | wc -l''', returnStdout: true).trim()
+                checkout scm
+                fix_result = sh(script: '''docker run --pull=always --name="$BUILD_TAG-ruff-fix" -e GIT_SRC="https://github.com/eea/$GIT_NAME.git" -e GIT_NAME="$GIT_NAME" -e GIT_BRANCH="$BRANCH_NAME" -e GIT_CHANGE_ID="$CHANGE_ID" eeacms/ruff check''', returnStatus: true)
+                sh '''docker cp $BUILD_TAG-ruff-fix:/code/$GIT_NAME .'''
+                sh '''cp -rf eea.api.dataconnector/* .'''
+                sh '''rm -rf eea.api.dataconnector'''
+                sh '''docker rm -v $BUILD_TAG-ruff-fix'''
+                FOUND_FIX = sh(script: '''git diff --name-only '*.py' | wc -l''', returnStdout: true).trim()
 
-                  if (FOUND_FIX != '0') {
-                    withCredentials([string(credentialsId: 'eea-jenkins-token', variable: 'GITHUB_TOKEN')]) {
-                      sh '''sed -i "s|url = .*|url = https://eea-jenkins:$GITHUB_TOKEN@github.com/eea/$GIT_NAME.git|" .git/config'''
-                    }
-                    sh '''git fetch origin $GIT_BRANCH:$GIT_BRANCH'''
-                    sh '''git checkout $GIT_BRANCH'''
-                    sh '''git add -- '*.py' '''
-                    sh '''git commit -m "lint: Automated code fix" '''
-                    sh '''git push --set-upstream origin $GIT_BRANCH'''
-                    sh '''exit 1'''
+                if (FOUND_FIX != '0') {
+                  withCredentials([string(credentialsId: 'eea-jenkins-token', variable: 'GITHUB_TOKEN')]) {
+                    sh '''sed -i "s|url = .*|url = https://eea-jenkins:$GITHUB_TOKEN@github.com/eea/$GIT_NAME.git|" .git/config'''
                   }
+                  sh '''git fetch origin $GIT_BRANCH:$GIT_BRANCH'''
+                  sh '''git checkout $GIT_BRANCH'''
+                  sh '''git add -- '*.py' '''
+                  sh '''git commit -m "lint: Automated code fix" '''
+                  sh '''git push --set-upstream origin $GIT_BRANCH'''
+                  sh '''exit 1'''
                 }
               }
             }
